# src/swagger.yaml
openapi: 3.0.3
info:
  title: MedVehicule API
  description: API REST de gestion de flotte de véhicules hospitaliers
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Serveur local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/vehicle-types:
    get:
      summary: Liste des types de véhicules
      tags: [Types de véhicules]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Liste }
    post:
      summary: Créer un type
      tags: [Types de véhicules]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        '201': { description: Créé }
  /api/vehicle-types/{id}:
    get:
      summary: Détail type
      tags: [Types de véhicules]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    patch:
      summary: Modifier type
      tags: [Types de véhicules]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        '200': { description: OK }
  /api/vehicles:
    get:
      summary: Liste des véhicules
      tags: [Véhicules]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: hospitalId
          in: query
          schema: { type: string }
        - name: vehicleTypeId
          in: query
          schema: { type: string }
        - name: available
          in: query
          schema: { type: string, enum: [true, false] }
      responses:
        '200': { description: Liste }
    post:
      summary: Créer un véhicule
      tags: [Véhicules]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [hospitalId, vehicleTypeId]
              properties:
                hospitalId: { type: string }
                vehicleTypeId: { type: string }
                registration: { type: string }
                brand: { type: string }
                model: { type: string }
                year: { type: integer }
      responses:
        '201': { description: Créé }
  /api/vehicles/{vehicleId}:
    get:
      summary: Détail véhicule
      tags: [Véhicules]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    patch:
      summary: Modifier véhicule
      tags: [Véhicules]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                registration: { type: string }
                brand: { type: string }
                model: { type: string }
                year: { type: integer }
                vehicleTypeId: { type: string }
      responses:
        '200': { description: OK }
    delete:
      summary: Supprimer véhicule
      tags: [Véhicules]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Supprimé }
...
  /api/users:
    get:
      summary: Liste des utilisateurs
      tags: [Utilisateurs]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Liste }
    post:
      summary: Créer un utilisateur
      tags: [Utilisateurs]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [hospitalId, email, password, firstName, lastName, role]
              properties:
                hospitalId: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
                firstName: { type: string }
                lastName: { type: string }
                role: { type: string, enum: [staff, supervisor, admin] }
      responses:
        '201': { description: Créé }
...
  /api/authorizations/user/{userId}:
    get:
      summary: Droits d'un utilisateur
      tags: [Autorisations]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/photos/vehicle/{vehicleId}:
    get:
      summary: Photos d'un véhicule
      tags: [Photos]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Liste }
    post:
      summary: Ajouter une photo
      description: multipart/form-data - photo (fichier), label, bookingId
      tags: [Photos]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo: { type: string, format: binary }
                label: { type: string }
                bookingId: { type: string }
      responses:
        '201': { description: Créé }
...
  /api/problem-reports:
    get:
      summary: Liste des signalements
      tags: [Signalements]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [ouvert, en_cours, cloture] }
      responses:
        '200': { description: Liste }
    post:
      summary: Créer un signalement
      tags: [Signalements]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [vehicleId, type]
              properties:
                vehicleId: { type: string }
                type: { type: string, enum: [rayure, impact, panne, autre] }
                description: { type: string }
                severity: { type: string, enum: [mineur, moyen, majeur] }
                bookingId: { type: string }
                photoIds: { type: array, items: { type: string } }
      responses:
        '201': { description: Créé }
  /api/problem-reports/{id}:
    get:
      summary: Détail signalement
      tags: [Signalements]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    patch:
      summary: Modifier signalement
      tags: [Signalements]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [ouvert, en_cours, cloture] }
                description: { type: string }
      responses:
        '200': { description: OK }
  /api/odometer/vehicle/{vehicleId}:
    get:
      summary: Relevés compteur
      tags: [Compteur]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        '200': { description: Liste }
    post:
      summary: Enregistrer un relevé
      tags: [Compteur]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: { type: integer, minimum: 0 }
                bookingId: { type: string }
      responses:
        '201': { description: Créé }
...
  /api/maintenance/contracts/vehicle/{vehicleId}:
    get:
      summary: Contrats d'un véhicule
      tags: [Entretien]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Liste }
...
  /api/vehicle3d/vehicle/{vehicleId}:
    get:
      summary: Modèle 3D du véhicule
      description: Retourne format et URL du fichier glTF/GLB (iOS, Android, Web)
      tags: [Modèle 3D]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: Modèle non disponible }
    post:
      summary: Upload modèle 3D
      description: multipart/form-data - model (fichier .glb ou .gltf)
      tags: [Modèle 3D]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                model: { type: string, format: binary }
      responses:
        '201': { description: Créé }
    delete:
      summary: Supprimer le modèle 3D
      tags: [Modèle 3D]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Supprimé }